Edited whole application 2

+++ app/src/main/java/com/example/familybudget/network/JoinGroupRequest.kt
@@ -2,5 +2,5 @@ package com.example.familybudget.network
 
 data class JoinGroupRequest(
     val groupId: Long,
-    val userId: Int
+    val userId: Long
 )
